ml-server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
psycopg2-binary==2.9.10
redis==5.2.1
sentence-transformers==3.4.1
huggingface-hub==0.29.3
torch==2.6.0
transformers==4.50.0
numpy==2.0.2
scikit-learn==1.6.1
pandas==2.2.3
aiohttp==3.11.4
dotenv==0.9.9